feat: add rotate functions
This will work in every case, but it does not check that the rotation
is right from a boolean logic point of view. It will rotate negation
operands (which will make the negation invalid), it will also try on
variables but it will do nothing since they dont have any operand.

It should only be done on commutative operations that have two operands
(so on conjunctions, disjunctions, exclusive disjunctions and logical
equivalences).